refactor(reddit_utils.py-,-reddit_post_scraper.py): refactored the reddit post scraper code
moved helper functions to reddit_utils and cleaned the main reddit contetn scraper code
