+++ reddit_post_scraper.py
@@ -7,6 +7,7 @@ import re
 import os
 from dotenv import load_dotenv
 from pprint import pprint
+from utils.reddit_utils import *
 
 
 # load env vars
@@ -38,46 +39,11 @@ def fetch_reddit_content(keywords,subreddits_ls):
     Args:
         subreddits_ls (List): List containing the subreddits
     """
-    def fetch_posts_url(query,sort,time_filter,subreddit,limit):
-        search_params = RedditSearchSchema(
-            query=query, sort=sort, time_filter=time_filter, subreddit=subreddit, limit=limit
-        )
-
-
-        result = search.run(tool_input=search_params.model_dump())
-        print(result)
-        url_pattern = r'https?://[^\s]+'
-        urls = re.findall(url_pattern, result)
-        return urls 
-
-    # https://praw.readthedocs.io/en/stable/tutorials/comments.html
-    # # print(reddit.user.me())
 
     all_comments = ''
-
-    def extract_comments(comment, level=0):
-        if level == 0:
-            header = "Comment:"
-        else:
-            header = "Reply:"
-        
-        indent = '    ' * level  # Indentation for nested comments
-        if level == 0:
-            formatted_comment = f"{indent}{header}\n{indent}{comment.body}\n\n"
-        else:
-            formatted_comment = f"{indent}{header}{level}\n{indent}{comment.body}\n\n"
-        
-        # Ensure all replies are loaded
-        comment.replies.replace_more(limit=None)
-        
-        for reply in comment.replies:
-            formatted_comment += extract_comments(reply, level + 1)
-        
-        return formatted_comment
-
-
     keywords_str = ' '.join(keywords)
     print(keywords_str)
+
     for subreddit in subreddits_ls:
         urls = fetch_posts_url(keywords_str,'relevance','year',subreddit,'1')
         print('Fetched urls')
@@ -92,12 +58,10 @@ def fetch_reddit_content(keywords,subreddits_ls):
                 for top_level_comment in submission.comments:
                     all_comments += extract_comments(top_level_comment)
                     all_comments += '-----------------------------\n'
-                # all_comments += fetch_comments(url) + '\n\n' + '-----------------------------------------------------------------------------------------'
     
     return all_comments
 
 
 comments  = fetch_reddit_content(['phuket'],['travelhacks'])
-# final_info  = '\n'.join(comments)
 print(comments)
